requirements.txt
Flask==0.10
Pillow==2.3
numpy==1.8
